requirements.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
Mako==1.1.6
MarkupSafe==2.0.1
nose==1.3.7
psycopg2-binary==2.9.2
PyJWT==1.4.2
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
Werkzeug==2.0.2
